my-vapi-tools/requirements.txt
flask==3.0.0
requests==2.32.4
python-dotenv==1.1.1
gunicorn==21.2.0
